+++ src/compile_fail/mod.rs
@@ -1,3 +1,0 @@
-// These modules contain `compile_fail` doc tests.
-
-mod unpin_conflict;